+++ tests/test_binary_sensor.py
@@ -99,6 +99,26 @@ async def test_binary_sensors(
     assert hot_water_tank_hot_water_flow_2.state == STATE_ON
     assert hot_water_tank_hot_water_flow_2.attributes[ATTR_FRIENDLY_NAME] == "Hot water tank 2 Hot water flow"
 
+    hot_water_tank_circulation_pump_state_1: State | None = hass.states.get(
+        "binary_sensor.keba_keenergy_12345678_hot_water_tank_circulation_pump_state_1",
+    )
+    assert isinstance(hot_water_tank_circulation_pump_state_1, State)
+    assert hot_water_tank_circulation_pump_state_1.state == STATE_ON
+    assert (
+        hot_water_tank_circulation_pump_state_1.attributes[ATTR_FRIENDLY_NAME]
+        == "Hot water tank 1 Circulation pump state"
+    )
+
+    hot_water_tank_circulation_pump_state_2: State | None = hass.states.get(
+        "binary_sensor.keba_keenergy_12345678_hot_water_tank_circulation_pump_state_2",
+    )
+    assert isinstance(hot_water_tank_circulation_pump_state_2, State)
+    assert hot_water_tank_circulation_pump_state_2.state == STATE_OFF
+    assert (
+        hot_water_tank_circulation_pump_state_2.attributes[ATTR_FRIENDLY_NAME]
+        == "Hot water tank 2 Circulation pump state"
+    )
+
     heat_pump_heat_request: State | None = hass.states.get(
         "binary_sensor.keba_keenergy_12345678_heat_pump_heat_request",
     )
@@ -231,6 +251,24 @@ async def test_binary_sensors_translations(
     )
     assert isinstance(hot_water_tank_hot_water_flow_2, State)
     assert hot_water_tank_hot_water_flow_2.attributes[ATTR_FRIENDLY_NAME] == "Warmwasserspeicher 2 Warmwasser läuft"
+
+    hot_water_tank_circulation_pump_state_1: State | None = hass.states.get(
+        "binary_sensor.keba_keenergy_12345678_hot_water_tank_circulation_pump_state_1",
+    )
+    assert isinstance(hot_water_tank_circulation_pump_state_1, State)
+    assert (
+        hot_water_tank_circulation_pump_state_1.attributes[ATTR_FRIENDLY_NAME]
+        == "Warmwasserspeicher 1 Zirkulationspumpe Status"
+    )
+
+    hot_water_tank_circulation_pump_state_2: State | None = hass.states.get(
+        "binary_sensor.keba_keenergy_12345678_hot_water_tank_circulation_pump_state_2",
+    )
+    assert isinstance(hot_water_tank_circulation_pump_state_2, State)
+    assert (
+        hot_water_tank_circulation_pump_state_2.attributes[ATTR_FRIENDLY_NAME]
+        == "Warmwasserspeicher 2 Zirkulationspumpe Status"
+    )
 
     heat_pump_heat_request: State | None = hass.states.get(
         "binary_sensor.keba_keenergy_12345678_heat_pump_heat_request",